# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agenda Inteligente — agenda-service",
...
    "version": "0.4.0"
  },
  "servers": [
    {
      "url": "https://cpdf-agenda.better-knowledge.com",
      "description": "Protótipo do programa"
    }
  ],
  "paths": {
...
    "/appointments/recorrentes": {
      "post": {
        "tags": [
          "recorrência"
        ],
        "summary": "Cria uma série recorrente ('toda terça às 10h até dezembro')",
        "description": "**Escopo:** `agenda:write`\n\nSérie semanal ou quinzenal a partir de `inicio` (primeira ocorrência), com `ocorrencias` OU `fim_em`. Cada ocorrência vira um compromisso próprio com a mesma régua de lembretes. Ocorrência que cair em horário ocupado NÃO quebra a série: volta em `conflitos`, já com as 3 alternativas — ofereça-as ao cliente e agende com POST /appointments. Aceita Idempotency-Key.",
        "operationId": "criar_serie_appointments_recorrentes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurrenceIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurrenceOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "400": {
            "description": "Pedido malformado",
            "content": {
              "application/json": {
                "examples": {
                  "DATA_SEM_FUSO": {
                    "summary": "Horário sem offset de fuso",
                    "value": {
                      "code": "DATA_SEM_FUSO",
                      "message": "'inicio' veio sem offset de fuso horário.",
                      "hint": "Envie ISO 8601 com offset explícito, ex.: 2026-08-27T15:30:00-03:00.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:write",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Repetir a chamada com a mesma chave devolve a resposta original sem duplicar efeito. Obrigatório para agentes; use um UUID novo por intenção."
          }
        ]
      }
    },
    "/appointments/recorrentes/{series_id}/cancel": {
      "post": {
        "tags": [
          "recorrência"
        ],
        "summary": "Cancela TODAS as ocorrências futuras da série",
        "description": "**Escopo:** `agenda:cancel + agenda:operacao`\n\nDistinto de cancelar UMA ocorrência (POST /appointments/{id}/cancel — RF-06). Exige agenda:cancel. Disparado por agente, exige confirmação humana: a primeira chamada devolve 409 CONFIRMACAO_NECESSARIA com prévia e token. Ocorrências passadas ou já realizadas não mudam; os slots futuros voltam para a grade na hora. Exige também `agenda:operacao`: a série é entidade da operação, e cancelá-la inteira não é ato de atendimento.",
        "operationId": "cancelar_serie_appointments_recorrentes__series_id__cancel_post",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Repetir a chamada com a mesma chave devolve a resposta original sem duplicar efeito. Obrigatório para agentes; use um UUID novo por intenção."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesCancelIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerieCanceladaOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito — o hint diz como resolver",
            "content": {
              "application/json": {
                "examples": {
                  "CONFIRMACAO_NECESSARIA": {
                    "summary": "Ação irreversível: confirme com o humano e repita com o token",
                    "value": {
                      "code": "CONFIRMACAO_NECESSARIA",
                      "message": "Cancelamento é irreversível e exige confirmação humana.",
                      "hint": "Mostre a prévia ao humano e, com o OK, repita esta chamada com o confirmation_token do payload.",
                      "retryable": false,
                      "previa": {
                        "compromisso": "0b6ff65e-…",
                        "cliente": "Paula Andrade",
                        "horario": "quinta, 27 de agosto, 15h30"
                      },
                      "confirmation_token": "cancel.0b6ff65e….1756312200.f3ab…"
                    }
                  },
                  "CONFIRMACAO_INVALIDA": {
                    "summary": "confirmation_token não vale para esta ação",
                    "value": {
                      "code": "CONFIRMACAO_INVALIDA",
                      "message": "confirmation_token inválido para esta ação.",
                      "hint": "Refaça a chamada sem token para receber uma nova prévia e um token novo.",
                      "retryable": false
                    }
                  },
                  "CONFIRMACAO_EXPIRADA": {
                    "summary": "confirmation_token passou dos 5 minutos",
                    "value": {
                      "code": "CONFIRMACAO_EXPIRADA",
                      "message": "O confirmation_token expirou (validade de 5 minutos).",
                      "hint": "Refaça a chamada sem token, confirme com o humano e use o token novo.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:cancel + agenda:operacao"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alternativa": {
        "properties": {
          "inicio": {
            "type": "string",
            "title": "Inicio",
            "description": "ISO 8601 com offset"
          },
          "label_humano": {
            "type": "string",
            "title": "Label Humano"
          }
        },
        "type": "object",
        "required": [
          "inicio",
          "label_humano"
        ],
        "title": "Alternativa"
      },
...
      "AppointmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "resource_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resource Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "cliente_telefone": {
            "type": "string",
            "title": "Cliente Telefone"
          },
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          },
          "fim": {
            "type": "string",
            "format": "date-time",
            "title": "Fim"
          },
          "label_humano": {
            "type": "string",
            "title": "Label Humano",
            "description": "Pronto para falar com o cliente, pt-BR"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "agendado | confirmado | cancelado | realizado | no_show"
          },
          "origem": {
            "type": "string",
            "title": "Origem"
          },
          "risco_no_show": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risco No Show",
            "description": "baixo | medio | alto (IA-03)"
          },
          "risco_detalhe": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risco Detalhe",
            "description": "Como o risco foi somado: pontos, fatores e explicação (IA-03 é auditável)"
          },
          "observacoes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacoes"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id",
            "description": "Presente quando faz parte de uma série (RF-15)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "service_id",
          "resource_id",
          "cliente_nome",
          "cliente_telefone",
          "inicio",
          "fim",
          "label_humano",
          "status",
          "origem"
        ],
        "title": "AppointmentOut",
        "examples": [
          {
            "cliente_nome": "Paula Andrade",
            "cliente_telefone": "+5511998765432",
            "fim": "2026-08-27T16:30:00-03:00",
            "id": "0b6ff65e-4f2a-4c8d-9e1b-3a5c7d9f0e2a",
            "inicio": "2026-08-27T15:30:00-03:00",
            "label_humano": "quinta, 27 de agosto, 15h30",
            "origem": "agente",
            "resource_id": "6f1e0c2a-8f6e-4b9e-9d3a-0d5b2a7c9c02",
            "service_id": "b3f0a1d4-2c5e-4f6a-8b7c-9d0e1f2a3b4c",
            "status": "agendado"
          }
        ]
      },
...
      "ConflitoOcorrencia": {
        "properties": {
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          },
          "label_humano": {
            "type": "string",
            "title": "Label Humano"
          },
          "alternativas": {
            "items": {
              "$ref": "#/components/schemas/Alternativa"
            },
            "type": "array",
            "title": "Alternativas",
            "description": "Ofereça-as ao cliente e agende a ocorrência com POST /appointments"
          }
        },
        "type": "object",
        "required": [
          "inicio",
          "label_humano",
          "alternativas"
        ],
        "title": "ConflitoOcorrencia",
        "description": "Ocorrência que caiu em slot ocupado: a série não quebra (RF-15) —\nfica pendente com as 3 alternativas já propostas."
      },
      "ErroOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Código estável, ex.: SLOT_INDISPONIVEL"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "O que aconteceu, em uma frase"
          },
          "hint": {
            "type": "string",
            "title": "Hint",
            "description": "O próximo passo, escrito para o agente executar"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "description": "true = repetir a mesma chamada pode funcionar"
          },
          "alternativas": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Alternativa"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternativas",
            "description": "Só em SLOT_INDISPONIVEL: os 3 horários livres mais próximos"
          },
          "previa": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previa",
            "description": "Só em CONFIRMACAO_NECESSARIA: o que será feito, para mostrar ao humano"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token",
            "description": "Só em CONFIRMACAO_NECESSARIA: repita a chamada com ele após o OK humano (expira em 5 min)"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "hint",
          "retryable"
        ],
        "title": "ErroOut",
        "description": "Todo erro da API tem este formato. O `hint` é escrito para o agente\nagir — quando possível, a saída já vem no payload (ex.: alternativas)."
      },
...
      "RecurrenceIn": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          },
          "cliente_nome": {
            "type": "string",
            "minLength": 1,
            "title": "Cliente Nome"
          },
          "cliente_telefone": {
            "type": "string",
            "minLength": 3,
            "title": "Cliente Telefone",
            "description": "E.164 no WhatsApp ou tg:<chat_id> no Telegram"
          },
          "frequencia": {
            "type": "string",
            "title": "Frequencia",
            "description": "semanal ou quinzenal"
          },
          "ocorrencias": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 52.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocorrencias"
          },
          "fim_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fim Em"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "origem": {
            "type": "string",
            "title": "Origem",
            "default": "agente"
          },
          "observacoes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacoes"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "inicio",
          "cliente_nome",
          "cliente_telefone",
          "frequencia"
        ],
        "title": "RecurrenceIn",
        "description": "RF-15: série semanal ou quinzenal — deliberadamente sem RRULE completo.\n\n`inicio` é a primeira ocorrência (define dia da semana e hora). Informe\n`ocorrencias` OU `fim_em`, nunca os dois.",
        "examples": [
          {
            "cliente_nome": "Paula Andrade",
            "cliente_telefone": "+5511998765432",
            "frequencia": "semanal",
            "inicio": "2026-09-08T10:00:00-03:00",
            "ocorrencias": 4,
            "service_id": "b3f0a1d4-2c5e-4f6a-8b7c-9d0e1f2a3b4c"
          }
        ]
      },
      "RecurrenceOut": {
        "properties": {
          "series_id": {
            "type": "string",
            "format": "uuid",
            "title": "Series Id"
          },
          "frequencia": {
            "type": "string",
            "title": "Frequencia"
          },
          "criadas": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "type": "array",
            "title": "Criadas"
          },
          "conflitos": {
            "items": {
              "$ref": "#/components/schemas/ConflitoOcorrencia"
            },
            "type": "array",
            "title": "Conflitos",
            "description": "Ocorrências que caíram em horário ocupado — pendentes, com alternativas"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "frequencia",
          "criadas",
          "conflitos"
        ],
        "title": "RecurrenceOut"
      },
...
      "SerieCanceladaOut": {
        "properties": {
          "series_id": {
            "type": "string",
            "format": "uuid",
            "title": "Series Id"
          },
          "canceladas": {
            "type": "integer",
            "title": "Canceladas",
            "description": "Quantas ocorrências futuras foram canceladas"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "canceladas"
        ],
        "title": "SerieCanceladaOut"
      },
      "SeriesCancelIn": {
        "properties": {
          "motivo": {
            "type": "string",
            "title": "Motivo"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token",
            "description": "Exigido para agentes: venha da resposta 409 CONFIRMACAO_NECESSARIA"
          }
        },
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "SeriesCancelIn",
        "examples": [
          {
            "motivo": "cliente encerrou o pacote"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT do Supabase Auth com claim `org_id` — humanos e UI."
      },
      "AgentKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key",
        "description": "Chave estática de agente (fase 1 do conector — vira OAuth 2.1 no agenda-mcp). Ações irreversíveis pedem confirmação humana."
      }
    }
  },
...
  "security": [
    {
      "SupabaseJWT": []
    },
    {
      "AgentKey": []
    }
  ]
}